Removing a few CI pipelines
Github Actions added to pre-compile artifacts for flang

Github Actions to build release_11x and release_12x

release_100 is still built as it was.
Artifacts names include the branch name.

Enable ccache for GitHub Actions

Allow full path to flang1 and flang2 in classic flang tests.

Remove unnecessary caret from path in classic-flang test

Remove the builds for gcc-9 and llvm-9

Build and test Flang project on PRs

Prepare llvm and clang for ARM64 build of Flang

Run Dockerfile script on Github's server to prepare a docker image with llvm and clang built and installed for Flang to use when testing code.