Fix -fveclib=PGMATH
The use of #ifdef in include/clang/Driver/Options.td was incorrect and
unsupported. As a result -fveclib=PGMATH was silently ignored, and
in LLVM 12, it causes the invocation to fail. This patch unguards the
option so that it is parsed correctly, but lets the FLANG_LLVM_EXTENSIONS
macro continue to toggle the feature.