Canonicalize some ENABLE_CLASSIC_FLANG
These CMake variables are used to set python variables in the lit
config. As such, they need to bet set with a valid CMake boolean value
that is also a valid python boolean value (1, True, etc.) and not one
that is not (ON, NO, etc.) to work. This is fragile.

Call the LLVM cmake helper function on these two downstream CMake