[build] Stop caching AArch64 Docker image layers
The "Pre-compile llvm ARM64" workflow tries to cache Docker image layers
to avoid rebuilding the same content over and over, as PRs are merged
into the release branches. Mysteriously, in older release branches the
workflow was not successfully caching layers, while in release_14x, it is
too successful, resulting in the same image being uploaded for successive
commit hashes. The actual build process is not too time-consuming (about
15 minutes) so it is safer and simpler to turn off caching completely.