[Driver] Pass Classic Flang libraries to the linker correctly
In LLVM 15, the Fortran and OpenMP runtime libraries are added to the linker
command line using common methods (addFortranRuntime* and addOpenMPRuntime*).
This commit adds Classic Flang awareness to addFortranRuntime*, so that
Classic Flang doesn't attempt to link with LLVM Flang libraries. Re-using
the same methods as Clang and LLVM Flang also helps reduce downstream delta.
A Classic Flang test is added to ensure that the linker command is constructed
correctly.