[workflows] Minimal workflow to create an AArch64 container
AArch64 GitHub workflows require a Docker container named after the branch.
This commit adds the minimal workflow required to create the container image
for building and testing subsequent pull requests.

Cherry-picking note: This commit also squashes parts of the following commits
in order to modernize the tools in the Docker image.

    e698fc84e984 [workflows] Update GitHub actions to use modern tools
    6aeab31be7f1 [workflows] Fix GitHub action bugs