[LinkerWrapper] Fix passing `-rpath` directly to clang
Summary:
This code passed the value of `-rpath` directly to the clang invocation.
If we're using the linker then it'll be fine. However, if the linker is
`gcc` as is the case when doing `-fopenmp-targets=x86_64` then this will
cause problems.  This patch adds the `-Wl,-rpath,` to feed it to the
linker correctly.

(cherry picked from 51ff5481146475afc869cf54ebc0b46d9da15a14)