[Driver][ClassicFlang] Add options -fno-automatic and -f(no-)implicit-none
These three options are added in release_12. When compiling programs
with them using Classic Flang, there are only warnings generated. They
are removed now so that the errors of "unknown argument" are reported.
Add these options in Driver for compatibility reason. Please note that
these options are not implemented yet in Classic Flang.