[Driver] Fix flang driver preprocessor issue
The -save-temps option has been causing the driver to run flang1/flang2
on the input file twice, and the fortran-preprocessor.f90 test was unable
to detect the error, because it had used -c instead of -x in a couple of
RUN commands.  This patch fixes the test and the driver to make sure that
the frontend is not invoked multiple times with -save-temps.