[workflows] Update GitHub actions to use modern tools
Cherry-pick patches from release_17x, in order to build the release_16x branch
of classic-flang-llvm-project and flang with GCC 12 and Clang 14/15
on Ubuntu 22.04. This squashes together the following commits:

    e698fc84 [workflows] Update GitHub actions to use modern tools
    6aeab31b [workflows] Fix GitHub action bugs
    77979af8 [workflows] Fix typo in pre-compile_llvm.yml; NFC

The workflows in the flang repository test both release_16x and release_17x,
and keeping the workflows consistent in these classic-flang-llvm-project
branches simplifies the flang workflows.