[workflows] Upgrade CMake in Ubuntu AArch64 container image
CMake 3.20 suffered from a bug that caused `find_library` to malfunction
on Ubuntu when Clang prepends `$CLANG_DIR/lib/$LLVM_HOST_TARGET` to the
search path for the linkder `ld`. This patch upgrades the version of
CMake used for the AArch64 CI workflow, to match the one used for the X86
CI workflow, which also fixes the bug.

Reference: https://gitlab.kitware.com/cmake/cmake/-/issues/25136